+++ ThreadsExperiment/ViewController.swift
@@ -52,15 +52,15 @@ class ViewController: UIViewController
         switch parameterButtonBar.selectedSegmentIndex
         {
             case 0:
-                parameters.f = Double(parameterSlider.value);
+                parameters.f = Float(parameterSlider.value);
             case 1:
-                parameters.k = Double(parameterSlider.value);
+                parameters.k = Float(parameterSlider.value);
             case 2:
-                parameters.dU = Double(parameterSlider.value);
+                parameters.dU = Float(parameterSlider.value);
             case 3:
-                parameters.dV = Double(parameterSlider.value);
+                parameters.dV = Float(parameterSlider.value);
             default:
-                parameters.f = Double(parameterSlider.value);
+                parameters.f = Float(parameterSlider.value);
         }
         
         updateLabel();
